requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-generativeai>=0.3.0
openai>=1.0.0
langchain>=0.0.340
langchain-google-genai>=0.0.6
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=23.2.0
razorpay>=1.3.0
watchfiles>=0.21.0
google-api-core
grpcio
grpcio-status
